use double for pose accumulation

+++ alley_oop/pose/raft_pose_estimator.py
@@ -24,7 +24,7 @@ class RAFTPoseEstimator(torch.nn.Module):
         model.eval()
         self.model = model
         self.intrinsics = intrinsics.unsqueeze(0).float()
-        self.last_pose= torch.nn.Parameter(init_pose.float(), requires_grad=False)
+        self.last_pose= torch.nn.Parameter(init_pose.double(), requires_grad=False)
         self.cost = [0]
         self.last_frame = None
         self.frame2frame = frame2frame
@@ -35,21 +35,21 @@ class RAFTPoseEstimator(torch.nn.Module):
             if self.last_frame is not None:
                 rel_pose_se3 = self.model(255*self.last_frame.img, 255*frame.img, self.intrinsics, self.baseline, depth1=self.last_frame.depth, depth2=frame.depth,
                                           mask1=self.last_frame.mask, mask2=frame.mask)[1].squeeze(0)
-                rel_pose = pseudo_lie_se3_to_SE3(rel_pose_se3)
+                rel_pose = pseudo_lie_se3_to_SE3(rel_pose_se3.double())
                 ret_frame = self.last_frame
             else:
-                rel_pose = torch.eye(4, dtype=torch.float32, device=self.last_pose.device)
+                rel_pose = torch.eye(4, dtype=torch.float64, device=self.last_pose.device)
                 ret_frame = None
             self.last_frame = frame
         else:
             # transform scene to last camera pose coordinates
-            scene_tlast = scene.transform_cpy(inv_transform(self.last_pose))
+            scene_tlast = scene.transform_cpy(inv_transform(self.last_pose.float()))
             model_frame = scene_tlast.render(self.intrinsics.squeeze())
             rel_pose_se3 = self.model(255*model_frame.img, 255*frame.img, self.intrinsics, self.baseline, depth1=model_frame.depth, depth2=frame.depth, mask1=model_frame.mask, mask2=frame.mask)[1].squeeze(0)
-            rel_pose = pseudo_lie_se3_to_SE3(rel_pose_se3)
+            rel_pose = pseudo_lie_se3_to_SE3(rel_pose_se3.double())
             ret_frame = model_frame
         self.last_pose.data = self.last_pose.data @ rel_pose
-        return self.last_pose, ret_frame
+        return self.last_pose.float(), ret_frame
 
     @property
     def device(self):
